+++ src/App.scss
@@ -1,12 +1,15 @@
-@import './design-tokens/scss-variables.scss';
+// @import './design-tokens/scss-variables.scss';
 
-.App-header {
-  background-color: var(--color-primary);
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
+// .App-header {
+//   background-color: var(--color-primary);
+//   min-height: 100vh;
+//   display: flex;
+//   flex-direction: column;
+//   align-items: center;
+//   justify-content: center;
+//   font-size: calc(10px + 2vmin);
+//   color: white;
+// }
+
+inner-column {
 }
