+++ src/screens/main/Intro/Intro.scss
@@ -1,0 +1,30 @@
+@import '../../../design-tokens/scss-variables.scss';
+
+main {
+  .contact {
+    display: flex;
+    justify-content: flex-end;
+    padding: $low-pad;
+  }
+}
+header {
+  font-family: $roboto;
+  padding: $low-pad;
+  h1 {
+    font-size: $attention;
+  }
+  h3 {
+    font-size: $calm;
+    color: var(--color-primary);
+    text-align: center;
+  }
+}
+
+.inner-column {
+  display: block;
+  width: 98%;
+  max-width: 1100px;
+  margin-right: auto;
+  margin-left: auto;
+  padding: $low-pad;
+}
