+++ src/screens/main/Intro/Intro.tsx
@@ -1,0 +1,20 @@
+import './Intro.scss';
+import React from 'react';
+
+function Intro() {
+  return (
+    <>
+      <main>
+        <div className="inner-column">
+          <section className="contact">(708) 477-5113</section>
+          <header className="welcome">
+            <h1>Welcome to SixKeys - Virtual Queue</h1>
+            <h3>"stop wasting valuable time"</h3>
+          </header>
+        </div>
+      </main>
+    </>
+  );
+}
+
+export default Intro;
